initial commit: planned features
navfs is a ls/cd/cp/mv/rm replacement.

This initial commit lays down the goal and vision of the project. as
speed and simplicity are the main goals of the project, it shouldn't go
beyond these initial planned features.

Once the application has enough functionality to be a daily driver,
it'll be rewritten from scratch and released as a 1.0.