feat(tui): basic tui and functionality done
The current goal is to implement as many features as possible to have a
better understanding of the structure of the application. once the
application has all of the initially planned features it'll be rewritten
from scracth.

By a loose interpretation of the SemVer quote "If your software is being
used in production, it should probably already be 1.0.0.", once the
application is good enough for me to use it as my main cd/ls/cp/mv/rm
replacement, I'll start rewriting the whole project to release it as
1.0.

Because the goal of the application is to be as minimal and fast as
possible, a 2.0 version will problably never happen.